backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
djoser
google-genai
llama-index-storage-chat-store-postgres
